Upgrade to golang 1.13.1
The change we care about: https://go-review.googlesource.com/c/go/+/114316/
We want to an HTTP POST to begin _immediately_ once the client has
started. Golang <=1.10 would buffer the request ... meaning the server
would have an open connection but would not receive HTTP headers until
the client's buffer is full.

In our case, the server garbage-collects any endpoint the client has not
used within 15s ... because it's expecting the client to connect
immediately and _then_ idle.

This came up when uploading a PST. The buffer wasn't being flushed until
30s in. I don't know why extract-pst took so long ... but let's solve
this bug first, then look into that one.

Code changes required: we need to change a FindSubmatch to
FindSubmatchIndex. We had been relying on golang's internals, and I
don't see a good reason for it. Let's hope I didn't break anything :).